Revert "Lockscreen is invisible when not needed to be visible"
This reverts commit 70cd49618fbffee313291ef2595f932f4bb2d85a.

Lockscreen visibility is not directly tied to LipstickSettings,
doing so breaks animation and makes for really terrible UX like
making the lockscreen vanish as soon as you lift your finger.

I'm aware that this is sometimes displayed when it shouldn't be
(like, for instance, when the homescreen initially launches),
but this fix is much worse.